add new input shapes

+++ data/slice_creator.py
@@ -8,11 +8,6 @@ from dasf_seismic.datasets import DatasetSEGY
 from dasf.datasets import DatasetArray, DatasetZarr
 
 
-def segy_reader(dataset_path):
-    dataset = DatasetSEGY("SEG-Y dataset", download=False, root=dataset_path)
-    return dataset.load()._data
-
-
 datasets = {"SEG-Y": DatasetSEGY, "Zarr": DatasetZarr, "NPY": DatasetArray}
 
 
@@ -32,14 +27,17 @@ def creator(args):
     for sli, sli_conf in dataset_configs["slices"].items():
         sli_size = list(map(int, sli.split("-")))
         for conf in sli_conf:
-            np.save(
-                os.path.join(args.name, sli, conf["name"]),
-                dataset[
-                    conf["start"][0] : conf["start"][0] + sli_size[0],
-                    conf["start"][1] : conf["start"][1] + sli_size[1],
-                    conf["start"][2] : conf["start"][2] + sli_size[2],
-                ],
-            )
+            if not os.path.exists(os.path.join(args.name, sli, conf["name"])):
+                if not os.path.exists(os.path.join(args.name, sli)):
+                    os.mkdir(os.path.join(args.name, sli))
+                np.save(
+                    os.path.join(args.name, sli, conf["name"]),
+                    dataset[
+                        conf["start"][0] : conf["start"][0] + sli_size[0],
+                        conf["start"][1] : conf["start"][1] + sli_size[1],
+                        conf["start"][2] : conf["start"][2] + sli_size[2],
+                    ],
+                )
 
 
 if __name__ == "__main__":
